requirements.txt
Flask==2.1.2
gunicorn==21.2.0
Werkzeug==2.3.3
Jinja2==3.1.2
MarkupSafe==2.1.5
Flask-CKEditor==0.4.6
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.25
WTForms==3.1.1
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9